refactor(package) : package remove and move
의미 없이 설계된 패키지 삭제하고
domain과 view로만 패키지 분리
