+++ src/main/java/com/woowahan/precourse/baseball/domain/BaseBallRecord.java
@@ -1,4 +1,4 @@
-package com.woowahan.precourse.baseball.record;
+package com.woowahan.precourse.baseball.domain;
 
 public class BaseBallRecord {
