+++ src/main/java/com/woowahan/precourse/baseball/domain/GameStatus.java
@@ -1,4 +1,4 @@
-package com.woowahan.precourse.baseball;
+package com.woowahan.precourse.baseball.domain;
 
 public enum GameStatus {
     RESTART("1"),
