+++ src/test/java/com/woowahan/precourse/baseball/record/BaseBallRecordTest.java
@@ -2,6 +2,7 @@ package com.woowahan.precourse.baseball.record;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.woowahan.precourse.baseball.domain.BaseBallRecord;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
