+++ src/test/java/com/woowahan/precourse/baseball/rule/BaseBallRuleTest.java
@@ -2,6 +2,7 @@ package com.woowahan.precourse.baseball.rule;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.woowahan.precourse.baseball.domain.BaseBallRule;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
